refactor(model): extract reusable functions into a new `FormatUtils` class
